PRO-2997: move subfield_contradictions.rb into core/contract/
File move only; the next commit changes the nesting.
