+++ lib/axn/core/contract/subfield_contradictions.rb
