+++ spec/axn/core/contract/subfield_contradictions_spec.rb
